refactor: move LinkFilesWithProviders out of UnrecognizedUtilityTabs
